refactor: configuración app.ts e index.ts
